Changed the directory layout of api.py, splitting each resource into their own file
